requirements.txt
Flask==3.0.3
selenium==4.19.0
mysql-connector-python==8.3.0
requests==2.31.0
beautifulsoup4==4.12.3